fix : customer resopnse 수정
- kakaoMap을 이용한 좌표값 반환 코드 추가

issue #112

+++ src/main/java/ga/backend/dong2/entity/Dong2.java
@@ -29,10 +29,10 @@ public class Dong2 extends Auditable {
     private String dongName;
 
     @Column
-    private Long x;
+    private Double x;
 
     @Column
-    private Long y;
+    private Double y;
 
     @Column
     private Boolean delYn = false;
